Move config and utilities to their own files to make setup easier
